move glob matching into its own file
